requirements.txt
pydrive2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
streamlit
pandas
requests
playwright
openpyxl
Pillow
plotly
